Safari-609.1.11.1/LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-drag-expected.txt

This test pauses the video, presses the fullscreen button, and drags the controls when the mouse is over a button. It should follow the mouse drag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!mediaControls.querySelector('.buttons-container.left') became true
PASS mediaControls.querySelector('.buttons-container.left').style.width became "118px"
PASS mediaControls.querySelector('.controls-bar').style.transform is "translate(-50px, -100px)"
PASS successfullyParsed is true

TEST COMPLETE
